Refactor PollResultsVC Section views to be more simple and flat hierarchy
